--- a/lib/lambda/planning.js
+++ b/lib/lambda/planning.js
@@ -549,6 +549,9 @@ const planning = {
     const plan = []
     let updatedScript = planning.ensureScriptGenesis(script, timeNow)
 
+    // Remove phases with a 0 second duration
+    updatedScript = planning.removeInvalidPhases(updatedScript)
+
     // ## Duration ##
     const scriptDurationInSeconds = planning.scriptDurationInSeconds(updatedScript)
 
@@ -595,6 +598,18 @@ const planning = {
     Object.assign(script, {
       _genesis: script._genesis === undefined ? timeNow : script._genesis,
     }),
+
+  removeInvalidPhases: (script) => {
+    const updatedScript = JSON.parse(JSON.stringify(script))
+
+    // Remove phases that have explicitly set a duration to 0
+    const newPhases = updatedScript.config.phases.filter(phase => !('duration' in phase) || phase.duration > 0)
+    const phasesWithDuration = updatedScript.config.phases.filter(phase => 'duration' in phase)
+
+    // Update the phases in the script. If there are no phases with a duration (e.g. only `pause` phases), then remove everything
+    updatedScript.config.phases = phasesWithDuration.length > 0 ? newPhases : []
+    return updatedScript
+  }
 }
 
 module.exports = planning