--- a/lib/lambda/planning.js
+++ b/lib/lambda/planning.js
@@ -609,7 +609,7 @@ const planning = {
     // Update the phases in the script. If there are no phases with a duration (e.g. only `pause` phases), then remove everything
     updatedScript.config.phases = phasesWithDuration.length > 0 ? newPhases : []
     return updatedScript
-  }
+  },
 }
 
 module.exports = planning