--- a/lib/lambda/planning.js
+++ b/lib/lambda/planning.js
@@ -549,9 +549,6 @@ const planning = {
     const plan = []
     let updatedScript = planning.ensureScriptGenesis(script, timeNow)
 
-    // Remove phases with a 0 second duration
-    updatedScript = planning.removeInvalidPhases(updatedScript)
-
     // ## Duration ##
     const scriptDurationInSeconds = planning.scriptDurationInSeconds(updatedScript)
 
@@ -575,7 +572,8 @@ const planning = {
       plan.push.apply(plan, parts) // eslint-disable-line prefer-spread
     }
 
-    return plan
+    // Remove phases with a 0 second duration
+    return planning.resolveInvalidPhases(plan)
   },
   /**
    * Plan for executing a set of samples for each unique flow in the given script where each sample
@@ -599,17 +597,57 @@ const planning = {
       _genesis: script._genesis === undefined ? timeNow : script._genesis,
     }),
 
-  removeInvalidPhases: (script) => {
-    const updatedScript = JSON.parse(JSON.stringify(script))
+  /**
+   * Given a `plan`, resolves any invalid phases.
+   * An invalid phase is defined as one of the following:
+   *   - Any phase with a duration of `0` following another phase that contains a duration. This phase will be removed.
+   *   - Any other phases with a duration of `0` (e.g. following a pause phase). These phases will be amended to have their duration set to `1`.
+   *
+   * @param plan An array of scripts to resolve
+   * @returns {Array} The resolved array of scripts with invalid phases amended
+   */
+  resolveInvalidPhases: (plan) => plan.map((script) => {
+    const newPhases = []
+    script.config.phases
+      .forEach((thisPhase, index) => {
+        const prevPhase = script.config.phases[index - 1]
+        const nextPhase = script.config.phases[index + 1]
+        if (thisPhase.duration > 0) {
+          newPhases.push(thisPhase)
+          return
+        }
 
-    // Remove phases that have explicitly set a duration to 0
-    const newPhases = updatedScript.config.phases.filter(phase => !('duration' in phase) || phase.duration > 0)
-    const phasesWithDuration = updatedScript.config.phases.filter(phase => 'duration' in phase)
+        // Remove any phases with a duration of 0 that follow a phase with a duration
+        // e.g.
+        // { duration: 1, arrivalRate: 0, rampTo: 25 }, { duration: 0, arrivalRate: 25 }
+        // Becomes
+        // { duration: 1, arrivalRate: 0, rampTo: 25 }
+        if (prevPhase && prevPhase.duration >= 0 && thisPhase.duration === 0) {
+          return
+        }
 
-    // Update the phases in the script. If there are no phases with a duration (e.g. only `pause` phases), then remove everything
-    updatedScript.config.phases = phasesWithDuration.length > 0 ? newPhases : []
-    return updatedScript
-  },
+        // Remove any pause phases that precede a phase with a duration of 0 - the duration phase will replace it
+        // Replace a previous pause phase if a duration of 0 follows. Set the phase duration to the pause duration
+        // e.g.
+        // { pause: 3 }, { duration: 0, arrivalRate: 1 }
+        // Becomes
+        // { duration: 3, arrivalRate: 1 }
+        if (nextPhase && nextPhase.duration === 0 && thisPhase.pause) {
+          return
+        }
+        if (prevPhase && prevPhase.pause && thisPhase.duration === 0) {
+          thisPhase.duration = prevPhase.pause // eslint-disable-line no-param-reassign
+          newPhases.push(thisPhase)
+          return
+        }
+
+        // Valid phase if we've made it here
+        newPhases.push(thisPhase)
+      })
+
+    script.config.phases = newPhases // eslint-disable-line no-param-reassign
+    return script
+  }),
 }
 
 module.exports = planning