--- a/lib/lambda/planning.js
+++ b/lib/lambda/planning.js
@@ -572,7 +572,6 @@ const planning = {
       plan.push.apply(plan, parts) // eslint-disable-line prefer-spread
     }
 
-    // Remove phases with a 0 second duration
     return planning.resolveInvalidPhases(plan)
   },
   /**
@@ -601,7 +600,8 @@ const planning = {
    * Given a `plan`, resolves any invalid phases.
    * An invalid phase is defined as one of the following:
    *   - Any phase with a duration of `0` following another phase that contains a duration. This phase will be removed.
-   *   - Any other phases with a duration of `0` (e.g. following a pause phase). These phases will be amended to have their duration set to `1`.
+   *   - Any other phases with a duration of `0` following a pause phase. The pause phase will be replaced by the duration phase
+   *     with a duration set equal to the pause duration.
    *
    * @param plan An array of scripts to resolve
    * @returns {Array} The resolved array of scripts with invalid phases amended