--- a/lib/lambda/planning.js
+++ b/lib/lambda/planning.js
@@ -608,42 +608,41 @@ const planning = {
    */
   resolveInvalidPhases: (plan) => plan.map((script) => {
     const newPhases = []
-    script.config.phases
-      .forEach((thisPhase, index) => {
-        const prevPhase = script.config.phases[index - 1]
-        const nextPhase = script.config.phases[index + 1]
-        if (thisPhase.duration > 0) {
-          newPhases.push(thisPhase)
-          return
-        }
-
-        // Remove any phases with a duration of 0 that follow a phase with a duration
-        // e.g.
-        // { duration: 1, arrivalRate: 0, rampTo: 25 }, { duration: 0, arrivalRate: 25 }
-        // Becomes
-        // { duration: 1, arrivalRate: 0, rampTo: 25 }
-        if (prevPhase && prevPhase.duration >= 0 && thisPhase.duration === 0) {
-          return
-        }
-
-        // Remove any pause phases that precede a phase with a duration of 0 - the duration phase will replace it
-        // Replace a previous pause phase if a duration of 0 follows. Set the phase duration to the pause duration
-        // e.g.
-        // { pause: 3 }, { duration: 0, arrivalRate: 1 }
-        // Becomes
-        // { duration: 3, arrivalRate: 1 }
-        if (nextPhase && nextPhase.duration === 0 && thisPhase.pause) {
-          return
-        }
-        if (prevPhase && prevPhase.pause && thisPhase.duration === 0) {
-          thisPhase.duration = prevPhase.pause // eslint-disable-line no-param-reassign
-          newPhases.push(thisPhase)
-          return
-        }
-
-        // Valid phase if we've made it here
+    script.config.phases.forEach((thisPhase, index) => {
+      const prevPhase = script.config.phases[index - 1]
+      const nextPhase = script.config.phases[index + 1]
+      if (thisPhase.duration > 0) {
         newPhases.push(thisPhase)
-      })
+        return
+      }
+
+      // Remove any phases with a duration of 0 that follow a phase with a duration
+      // e.g.
+      // { duration: 1, arrivalRate: 0, rampTo: 25 }, { duration: 0, arrivalRate: 25 }
+      // Becomes
+      // { duration: 1, arrivalRate: 0, rampTo: 25 }
+      if (prevPhase && prevPhase.duration >= 0 && thisPhase.duration === 0) {
+        return
+      }
+
+      // Remove any pause phases that precede a phase with a duration of 0 - the duration phase will replace it
+      // Replace a previous pause phase if a duration of 0 follows. Set the phase duration to the pause duration
+      // e.g.
+      // { pause: 3 }, { duration: 0, arrivalRate: 1 }
+      // Becomes
+      // { duration: 3, arrivalRate: 1 }
+      if (nextPhase && nextPhase.duration === 0 && thisPhase.pause) {
+        return
+      }
+      if (prevPhase && prevPhase.pause && thisPhase.duration === 0) {
+        thisPhase.duration = prevPhase.pause // eslint-disable-line no-param-reassign
+        newPhases.push(thisPhase)
+        return
+      }
+
+      // Valid phase if we've made it here
+      newPhases.push(thisPhase)
+    })
 
     script.config.phases = newPhases // eslint-disable-line no-param-reassign
     return script