--- a/lib/lambda/planning.js
+++ b/lib/lambda/planning.js
@@ -616,7 +616,7 @@ const planning = {
         return
       }
 
-      // Remove any phases with a duration of 0 that follow a phase with a duration
+      // Skip any phases with a duration of 0 that follow a phase with a duration
       // e.g.
       // { duration: 1, arrivalRate: 0, rampTo: 25 }, { duration: 0, arrivalRate: 25 }
       // Becomes
@@ -625,7 +625,7 @@ const planning = {
         return
       }
 
-      // Remove any pause phases that precede a phase with a duration of 0 - the duration phase will replace it
+      // Skip any pause phases that precede a phase with a duration of 0 - the duration phase will replace it
       // Replace a previous pause phase if a duration of 0 follows. Set the phase duration to the pause duration
       // e.g.
       // { pause: 3 }, { duration: 0, arrivalRate: 1 }