--- a/tests/integration/target/serverless.yml
+++ b/tests/integration/target/serverless.yml
@@ -22,6 +22,7 @@ functions:
 
   list:
     handler: handler.list
+    timeout: 30
     events:
       - http:
           path: list/{id}