--- a/tests/lib/lambda/planning.spec.js
+++ b/tests/lib/lambda/planning.spec.js
@@ -1256,6 +1256,9 @@ describe('./lib/lambda/taskPlan.js', () => {
         const expectedPhases2 = [
           { duration: 1, arrivalRate: 1 },
         ]
+        // NOTE: Prior to implementing the fix to resolve invalid phases, the `expectedPhases2` would have looked like:
+        // [ { pause: 1 }, { duration: 0, arrivalRate: 1 }]
+        // This would cause the lambda to run infinitely
         result = planning.planPerformance(1, script, defaultSettings)
         console.log(JSON.stringify(result, null, 2))
         expect(result.length).to.equal(2)