## Reason for change

The moment you've all been waiting for....
I present to you:

**Part 1 of a 2 part series called: FIXING THE DAMN PERFORMANCE TEST**

This part includes identification and resolution of a deeply buried serverless-artillery bug where a duration of `0` can occassionally be assigned to a load phase in some circumstances. This causes artillery to run it for an 'infinite' duration, timing out the lambda and this is what we see in the performance test error logs after a performance test has been run, partcularly for api-forecasts-town and api-observations.

I think it has something to do with specific durations that we specify in the performance test config, because it obviously only happens for a specific set of APIs, and other APIs operate perfectly fine.

For a more detailed explanation:
**UPDATE:** Upon further investigation, I found that the simplest configuration to cause this issue is:
```yml
duration: 1
arrivalRate: 0
rampTo: 26
```
The reason this happens is because the maximum RPS per lambda (chunk) is 25. When serverless-artillery tries to break this down, it performs a line intersection between the ramp phase: { x: 0, y: 0 ), { x: 1, y: 26 } and a limit: { x: 0, y: 25 }, { x: 1, y: 25 }
This returns an intersection point of (0.96..., 25), which is then rounded to (1, 25).
You will notice the the X value for the intersection point is equal to the duration of the phase. This is where the problem comes in, as the calculation performed to split the phase actually subtracts the x intersection point from the duration, resulting in a duration value of 0.


serverless-artillery restricts lambda invocations to 120s. Obviously, when we specify durations longer than this in our config (e.g 5min  ramp up from 0 to 3000 RPS), serverless-artillery needs to split the work up (and the load, because a single lambda cant do 3000 RPS). To achieve this, it recursively breaks down the phase until each 'chunk' is within pre-defined limits, set in lib/lambda/platform-settings.js.
Depending at what point in the 120s the lambda needs to execute it's load, serverless-artillery will add a series of `pause` phases. This is done recursively, so when you get to a lambda that is supposed to execute its load at the end of the 120s period, if you log the phases in that lambda you will see a whole bunch of 2 or 3 second `pause` phases before the actual load phase is defined.
The problem is, in some very rare cases, the `pause` phases (and sometimes a final ramp up phase) will exhaust the entire 120s of restricted runtime (this is just imposed by serverless-artillery, the lambda timeout is still 300s).
This can cause a final phase to be added which has a duration of `0`, because the amount left to ramp up or down to is 25 <= x < 30. When this happens, it seems that `artillery` doesn't really know what to do, and takes it as an infinite duration, causing the lambda to timeout.

The fix I have implemented is to remove any phases with a duration of 0 that follow a duration phase, and then replace a `pause` phase with the following duration phase if it has a duration of 0.
You can see some examples of this in the code comments, and I've added a test.

Unfortunately the original serverless-artillery repository has probably been private-d (it is a 404 now), and I was hoping an issue had been opened against this because it seems like it would be a fairly common issue - unless we've somehow managed to pick the few durations that recursively get reduced down to a phase with a 0 duration....

Ticket:

This is a:

- [X] Bug fix (non-breaking change which fixes an issue)
- [ ] Refactor (non-breaking change)
- [ ] New or improved tests (non-breaking change)
- [ ] New feature (non-breaking change which adds functionality or implements non-functional requirement)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)
- [ ] Documentation or design proposal


## Checklist

- [ ] This PR has a descriptive title
- [ ] My code follows the code style of this project.
- [ ] My change requires a change to the documentation and docs have been updated.
- [ ] I have added tests to cover my changes.
- [ ] All new and existing tests passed.
- [ ] This change requires an Akamai change.  Cherwell ticket number is in Reason for Change.
- [ ] This needs a change request.  CR# is in title.