@MichaelBellBOM 2022-11-20
While I support doing this now, it feels like we should consider shifting away from serverless-artillery as it looks like we'll have to provide maintenance for any issues with this.